+++ CHANGELOG.md
@@ -121,7 +121,8 @@
 - Enh #1016: Refactor `AbstractDMLQueryBuilder::getTableUniqueColumnNames()` method (@Tigrov)
 - New #1015: Add mode parameter to `Like` condition (@vjik)
 - Chg #1018: Split `BetweenColumns` condition to `BetweenColumns` and `NotBetweenColumns` (@vjik)
- 
+- Chg #1017: Split `Between` condition to `Between` and `NotBetween` (@vjik)
+
 ## 1.3.0 March 21, 2024
 
 - Enh #778: Deprecate unnecessary argument `$rawSql` of `AbstractCommand::internalExecute()` (@Tigrov)